requirements.txt
Django==5.1.2
django-ninja==1.3.0
PyJWT==2.9.0
pillow==11.0.0
email_validator==2.2.0
